Made timerDuration on Logic_Timer a float instead of an int

+++ CorridorGeodesic/Assets/Resources/Scripts/Integratables/Updated Logic System/Logic_Timer.cs
@@ -20,7 +20,7 @@ public class Logic_Timer: LogicComponent
     // Private Variables
     //=-----------------=
     [SerializeField] private LogicComponent inputSignal;
-    [SerializeField] private int timerDuration;
+    [SerializeField] private float timerDuration;
 
     //=-----------------=
     // Reference Variables
